+++ Mile.Helpers/Mile.Helpers.CppBase.h
@@ -38,6 +38,17 @@ namespace Mile
     std::wstring FormatWideString(
         _In_z_ _Printf_format_string_ wchar_t const* const Format,
         ...);
+
+    /**
+     * @brief Write formatted data to a onebyte or multibyte string, suggested
+     *        encoding with UTF-8.
+     * @param Format Format-control string.
+     * @param ArgList Pointer to list of optional arguments to be formatted.
+     * @return A formatted string if successful, an empty string otherwise.
+    */
+    std::string VFormatString(
+        _In_z_ _Printf_format_string_ char const* const Format,
+        _In_z_ _Printf_format_string_ va_list ArgList);
 }
 
 #endif // !MILE_WINDOWS_HELPERS_CPPBASE